% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionToViper.R
\name{addResponseClass}
\alias{addResponseClass}
\title{\code{addResponseClass} takes an expression set and thresholds data to call differential expression based on drug}
\usage{
addResponseClass(eset, drug, thresholds = c(0.25, 0.75))
}
\arguments{
\item{eset}{is expression set}

\item{thresholds}{are 2 values to use as quantiles}
}
\value{
expression set
}
\description{
\code{addResponseClass} takes an expression set and thresholds data to call differential expression based on drug
}
