% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFendR.R
\name{basicFendR}
\alias{basicFendR}
\title{An S3 class to represent a basic implementation of the fendR predictive
network algorithm}
\usage{
basicFendR(networkFile, featureData, phenoFeatureData, sampleOutcomeData,
  targetGenes, geneNorm, responseNorm)
}
\arguments{
\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{phenoFeatureData}{a data.frame where rows represent genes and columns represent a relationship between phenotype and gene}

\item{sampleOutcomeData}{a data.frame representing at least one column of phenotype and rows representing samples}
}
\value{
basicFendR object
}
\description{
An S3 class to represent a basic implementation of the fendR predictive
network algorithm
}
