% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{buildModelFromEngineeredFeatures.fendR}
\alias{buildModelFromEngineeredFeatures.fendR}
\title{Builds predictive model from network-augmented feature
\code{buildModelFromEngineeredFeatures} takes the engineered features and creates a model based on an underlying}
\usage{
\method{buildModelFromEngineeredFeatures}{fendR}(object, testDrugs = NULL)
}
\arguments{
\item{object}{That contains a phenotypic data}

\item{testDrug}{name of drug to select from dataset to test in smaller context}
}
\value{
an object of class \code{fendR} with a list of \code{lm} objects
}
\description{
Builds predictive model from network-augmented feature
\code{buildModelFromEngineeredFeatures} takes the engineered features and creates a model based on an underlying
}
