% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{buildModelFromEngineeredFeatures}
\alias{buildModelFromEngineeredFeatures}
\title{Builds predictive model from network-augmented feature
\code{buildModelFromEngineeredFeatures} takes the engineered features and creates a model based on an underlying}
\usage{
buildModelFromEngineeredFeatures(object)
}
\arguments{
\item{object}{That contains a phenotypic data}

\item{newFeatureSet}{from \code{createNewFeaturesFromNetwork}}
}
\value{
an object of class \code{lm}
}
\description{
Builds predictive model from network-augmented feature
\code{buildModelFromEngineeredFeatures} takes the engineered features and creates a model based on an underlying
}
