% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{buildModelFromOriginalFeatures.fendR}
\alias{buildModelFromOriginalFeatures.fendR}
\title{Builds predictive model from original features
\code{buildModelFromOriginalFeatures} takes the original features and creates a model based on an underlying}
\usage{
\method{buildModelFromOriginalFeatures}{fendR}(object, testDrugs = NA)
}
\arguments{
\item{object}{That contains a phenotypic data}

\item{testDrug}{name of drug to select from dataset to test in smaller context}
}
\value{
an fendR object with a list of \code{lm} objects
}
\description{
Builds predictive model from original features
\code{buildModelFromOriginalFeatures} takes the original features and creates a model based on an underlying
}
