% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{buildModelFromOriginalFeatures}
\alias{buildModelFromOriginalFeatures}
\title{Builds predictive model from original features
\code{buildModelFromOriginalFeatures} takes the original features and creates a model based on an underlying}
\usage{
buildModelFromOriginalFeatures(object, testDrugs)
}
\arguments{
\item{object}{That contains a phenotypic data}
}
\value{
an fendR object with a list of \code{lm} objects
}
\description{
Builds predictive model from original features
\code{buildModelFromOriginalFeatures} takes the original features and creates a model based on an underlying
}
