% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{calculate.auc}
\alias{calculate.auc}
\title{Calculate the Area Under the ROC Curve}
\usage{
calculate.auc(response, predicted.response, ...)
}
\arguments{
\item{response}{A vector of the response data}

\item{predicted.response}{A vector of the predicted response.}
}
\value{
The Area Under the ROC Curve
}
\description{
Calculate the Area Under the ROC Curve
}
