% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{calculate.roc}
\alias{calculate.roc}
\title{Calculate the ROC Curve}
\usage{
calculate.roc(response, predicted.response, ...)
}
\arguments{
\item{response}{A vector of the response data}

\item{predicted.response}{A vector of the predicted response.}
}
\value{
A list of class "roc"
}
\description{
Calculate the ROC Curve
}
