% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{cellLineMapping}
\alias{cellLineMapping}
\title{cellLineMapping}
\usage{
cellLineMapping()
}
\description{
\code{cellLineMapping} collects information from CCLE and SANGER datasets to ensure to create a table that maps cell lines from one to another
}
