% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{compareModelToRandom}
\alias{compareModelToRandom}
\title{Assess fendR and modeling using random weights}
\usage{
compareModelToRandom(fendRObj, testPheno)
}
\arguments{
\item{fendRObj}{The object class}

\item{testPheno}{a list of phenotypes to limit the test}
}
\value{
Not sure yet...
}
\description{
Compare performance of fendR with random weights added to network/features
}
\keyword{INCOMPLETE}
