% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFendR.R
\name{createNewFeaturesFromNetwork.basicFendR}
\alias{createNewFeaturesFromNetwork.basicFendR}
\title{Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the phenotype scores and updates the gene scores based on phenotype}
\usage{
\method{createNewFeaturesFromNetwork}{basicFendR}(object, testDrugs = NA,
  numCores = 1)
}
\arguments{
\item{object}{That contains a data frame and network}
}
\value{
list of gene features for each phenotype/drug response
}
\description{
Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the phenotype scores and updates the gene scores based on phenotype
}
