% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\name{createNewFeaturesFromNetwork.forestFendR}
\alias{createNewFeaturesFromNetwork.forestFendR}
\title{Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the gene-based measurements and alters their #' score using a network}
\usage{
\method{createNewFeaturesFromNetwork}{forestFendR}(object)
}
\arguments{
\item{object}{That contains a data frame and network}
}
\value{
data.frame representing new gene by sample matrix that is augmented
}
\description{
Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the gene-based measurements and alters their #' score using a network
}
