% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{createNewFeaturesFromNetwork}
\alias{createNewFeaturesFromNetwork}
\title{Engineer Features from Network}
\usage{
createNewFeaturesFromNetwork(object, ...)
}
\arguments{
\item{object}{fendR object after running \code{loadNetwork}}
}
\value{
fendR object with the remappedFeature data.frame populated with four columns: Gene, Sample, Phenotype, Value
}
\description{
Takes the igraph populated by \code{loadNetwork}
and propagates scores to remapped features
}
\keyword{network}
\keyword{propagation}
