% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{doAllNetworkAssess}
\alias{doAllNetworkAssess}
\title{\code{doAllNetworkAssess} does all the network stuff}
\usage{
doAllNetworkAssess(synId, tableId)
}
\arguments{
\item{synId}{synapse id of stored network result}

\item{tableId}{of stored fendR results}
}
\description{
\code{doAllNetworkAssess} does all the network stuff
}
