% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{drugNameMapping}
\alias{drugNameMapping}
\title{drugNameMapping}
\usage{
drugNameMapping()
}
\value{
a data frame with each drug and its aliass on each row
}
\description{
\code{drugNameMapping} collects information from the CCLE, SANGER and NTAP datasets to create a table that maps one drug to the others
}
