% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFendR.R
\name{engineeredResponseMatrix.basicFendR}
\alias{engineeredResponseMatrix.basicFendR}
\title{Get Engineered Features as model matrix}
\usage{
\method{engineeredResponseMatrix}{basicFendR}(fObj, phenotype = c())
}
\description{
Gets a \code{list} of response matrices for a phenotype
}
