% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3FendR.R
\name{engineeredResponseMatrix.n3FendR}
\alias{engineeredResponseMatrix.n3FendR}
\title{Return re-engineered feature matrix, consisting of re-engineered features and the associated responses in column Response.}
\usage{
\method{engineeredResponseMatrix}{n3FendR}(object, phenotype = c())
}
\value{
A matrix suitable for modeling with the formula Response ~ from the re-engineered features, whose columns are genes and the Response and whose rows are "phenotype" . "sample".
}
\description{
Creates a matrix suitable for modeling with the formula Response ~ from the re-engineered features.
}
\keyword{fendR}
\keyword{model}
