% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{engineeredResponseMatrix}
\alias{engineeredResponseMatrix}
\title{Get re-engineered feature matrix}
\usage{
engineeredResponseMatrix(object, phenotype = c())
}
\value{
A response matrix to use for modeling with the formula Response~.
}
\description{
Grabs data from the re-engineered features and creates a model-ready response matrix
}
\keyword{fendR}
\keyword{model}
