% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\docType{class}
\name{fendR-class}
\alias{fendR-class}
\alias{fendR}
\title{An S4 class to represent a fendR predictive network algorithm}
\description{
An S4 class to represent a fendR predictive network algorithm
}
\section{Slots}{

\describe{
\item{\code{network}}{A matrix (exact format TBD)}

\item{\code{featureData}}{a data.frame that contains rows representing genes and columns representing samples}

\item{\code{phenoData}}{a data.frame representing at least one column of phenotype and rows representing samples}
}}

