% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{fendR}
\alias{fendR}
\title{An S3 class to represent a fendR predictive network algorithm}
\usage{
fendR(network, featureData, phenoData)
}
\arguments{
\item{network}{A matrix (exact format TBD)}

\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{phenoData}{a data.frame representing at least one column of phenotype and rows representing samples}
}
\value{
a fendR object
}
\description{
An S3 class to represent a fendR predictive network algorithm
}
