% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\docType{class}
\name{forestFendr-class}
\alias{forestFendr-class}
\alias{forestFendR}
\title{An S4 class to represent a Forest-based implementation of the fendR predictive
network algorithm}
\description{
An S4 class to represent a Forest-based implementation of the fendR predictive
network algorithm
}
\section{Slots}{

\describe{
\item{\code{forestPath}}{Path to python forest scripts}
}}

