% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\name{forestFendR}
\alias{forestFendR}
\title{An S3 class to represent a Forest-based implementation of the fendR predictive
network algorithm}
\usage{
forestFendR(network, featureData, phenoData, forestPath)
}
\arguments{
\item{network}{A network or path to network, depending on the underlying class}

\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{forestPath}{Path to python forest scripts}
}
\value{
forestFendR object
}
\description{
An S3 class to represent a Forest-based implementation of the fendR predictive
network algorithm
}
