% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCSFWithDTEnetwork.R
\name{getDrugIds}
\alias{getDrugIds}
\title{\code{getDrugIds} Identifies drugs in a}
\usage{
getDrugIds(drug_names)
}
\arguments{
\item{pheno.file}{Tidied drug response}
}
\value{

}
\description{
\code{getDrugIds} Identifies drugs in a
}
