% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{getGraphSaveVis}
\alias{getGraphSaveVis}
\title{\code{getGraphSaveVis} Grabs a graph object stored on synapse, runs enrichment and saves visualiation}
\usage{
getGraphSaveVis(synId)
}
\arguments{
\item{synId}{synapse id of file}
}
\value{
not sure
}
\description{
\code{getGraphSaveVis} Grabs a graph object stored on synapse, runs enrichment and saves visualiation
}
