% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\name{getGraphStats}
\alias{getGraphStats}
\title{Get graph statistics for 'optimal' model selection
\code{getGraphStats} gets various statistics in network}
\usage{
getGraphStats(graphg)
}
\arguments{
\item{iGraph}{object}
}
\value{
list of statistics
}
\description{
Get graph statistics for 'optimal' model selection
\code{getGraphStats} gets various statistics in network
}
