% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionToViper.R
\name{getViperForDrug}
\alias{getViperForDrug}
\title{\code{getViperForDrug} takes an expression set and a list of drugs and identifies a viper list of proteins that explain differential expression}
\usage{
getViperForDrug(v.res, high, low, pvalthresh = 0.05, useEntrez = TRUE,
  p.corr = TRUE)
}
\arguments{
\item{drugs}{is a list of drug names}
}
\value{
viper object
}
\description{
\code{getViperForDrug} takes an expression set and a list of drugs and identifies a viper list of proteins that explain differential expression
}
