% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{getViperProtExpressionEnrich}
\alias{getViperProtExpressionEnrich}
\title{\code{getViperProtExpressionEnrich} Gets viper proteins from results and computes go BP enrichment}
\usage{
getViperProtExpressionEnrich(synId, tableId)
}
\arguments{
\item{synId}{synapse id of stored network result}

\item{tableId}{of stored fendR results}
}
\value{
data frame of all go enrichment with adjusted p-value less than 0.1
}
\description{
\code{getViperProtExpressionEnrich} Gets viper proteins from results and computes go BP enrichment
}
