% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionToViper.R
\name{loadEset}
\alias{loadEset}
\title{\code{loadEset} takes the rna seq and pheno test files and builds expression set}
\usage{
loadEset(rna.data, pheno.data, useEntrez = TRUE)
}
\arguments{
\item{rna.seq.data}{Tidied rna seq file}

\item{pheno.file}{Tidied drug response}
}
\value{
expression set
}
\description{
\code{loadEset} takes the rna seq and pheno test files and builds expression set
}
