% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFendR.R
\name{loadNetwork.basicFendR}
\alias{loadNetwork.basicFendR}
\title{Load Network data}
\usage{
\method{loadNetwork}{basicFendR}(fObj)
}
\arguments{
\item{Path}{to network file}
}
\value{
a fendR object with the graph parameter populated with an iGraph object where edge weights represent distance between nodes (smaller means *more* association)
}
\description{
\code{loadNetwork} takes a file path and formats it as network
}
\keyword{feather}
\keyword{network}
