% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{loadNetwork}
\alias{loadNetwork}
\title{Load Network data}
\usage{
loadNetwork(fname)
}
\arguments{
\item{Path}{to network file}
}
\value{
? matrix? data frame?
}
\description{
\code{loadNetwork} takes a file path and formats it as network
}
\keyword{feather}
\keyword{network}
