% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{loadPhenotypeData}
\alias{loadPhenotypeData}
\title{Load phenotype data}
\usage{
loadPhenotypeData(fname)
}
\arguments{
\item{Path}{to phenotype file}
}
\value{
tidied data frame with columns 'Sample','Phenotype','Response'
}
\description{
\code{loadPhenotypeData} loads in the phenotype data with the sample identifiers as rows and any phenotype of interest as the column
}
\keyword{phenotype}
