% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Load sample data}
\usage{
loadSampleData(fname)
}
\arguments{
\item{Path}{to sample file}
}
\value{
tidied data frame with columns 'Gene','Sample' and 'Value'
}
\description{
\code{loadSampleData} takes a file path and loads it into a data frame
}
