% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{loadTargetData}
\alias{loadTargetData}
\title{Load target/snp data}
\usage{
loadTargetData(fname)
}
\arguments{
\item{Path}{to  file}
}
\value{
Data frame with at least 2 column
}
\description{
\code{loadTargetData} loads in the target/phenotype data with the first column representing the drug of choice and the second represneting the gene
}
\keyword{phenotype}
