% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3FendR.R
\name{n3FendR}
\alias{n3FendR}
\title{An S3 class to represent the nearest-neighbor network implementation of the fendR predictive
network algorithm.}
\usage{
n3FendR(network, featureData, phenoFeatureData, sampleOutcomeData, targetGenes,
  testDrugs = NA)
}
\arguments{
\item{network}{the file name of a feather object representing a dense matrix}

\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{phenoFeatureData}{a data.frame where rows represent genes and columns represent a relationship between phenotype and gene}

\item{sampleOutcomeData}{a data.frame representing at least one column of phenotype and rows representing samples}

\item{testDrugs}{a vector of drug names to be used to subset the data (e.g., for debugging/efficiency purposes)}

\item{target.genes}{a vector of target gene names (a subset of those in featureData) that will be used to sparsify the network}
}
\value{
n3FendR object
}
\description{
An S3 class to represent the nearest-neighbor network implementation of the fendR predictive
network algorithm.
}
