% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3FendR.R
\name{originalResponseMatrix.n3FendR}
\alias{originalResponseMatrix.n3FendR}
\title{Return feature matrix, consisting of the original features and the associated responses in column Response.}
\usage{
\method{originalResponseMatrix}{n3FendR}(object, phenotype = c(),
  limit.to.engineered.genes = FALSE)
}
\value{
A matrix suitable for modeling with the formula Response ~ from the original features, whose columns are genes and the Response and whose rows are "phenotype" . "sample".
}
\description{
Creates a matrix suitable for modeling with the formula Response ~ from the original features.
}
\keyword{(which}
\keyword{Boolen}
\keyword{be}
\keyword{beyond}
\keyword{by}
\keyword{constructor}
\keyword{engineered}
\keyword{features}
\keyword{fendR}
\keyword{in}
\keyword{indicating}
\keyword{limit.to.engineered.genes}
\keyword{limited}
\keyword{may}
\keyword{model}
\keyword{network)}
\keyword{passed}
\keyword{restricted}
\keyword{returned}
\keyword{should}
\keyword{target.genes}
\keyword{the}
\keyword{those}
\keyword{to}
\keyword{used}
\keyword{whether}
