% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{originalResponseMatrix}
\alias{originalResponseMatrix}
\title{Get model-ready matrix from original features}
\usage{
originalResponseMatrix(object, phenotype = c(), ...)
}
\arguments{
\item{object}{a fendrObject}
}
\value{
A response matrix to use for modeling with the formula 'Response~.'
}
\description{
return the original feature data in a response matrix for modeling
}
