% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3FendR.R
\name{originalSparseResponseMatrix.n3FendR}
\alias{originalSparseResponseMatrix.n3FendR}
\title{Return feature matrix, consisting of the original features, and the associated responses in a separate data.frame.}
\usage{
\method{originalSparseResponseMatrix}{n3FendR}(object, phenotype = c(),
  limit.to.engineered.genes = FALSE)
}
\value{
A list with slots "feature.mat" (the feature matrix) and "response.mat" (a data.frame with columns "Sample", "Phenotype", and "Response").  The rows of feature.mat and response.mat are in 1-to-1 correspondence.
}
\description{
Creates a matrix suitable for modeling with the formula Response ~ from the original features.
}
\keyword{(which}
\keyword{Boolen}
\keyword{be}
\keyword{beyond}
\keyword{by}
\keyword{constructor}
\keyword{engineered}
\keyword{features}
\keyword{fendR}
\keyword{in}
\keyword{indicating}
\keyword{limit.to.engineered.genes}
\keyword{limited}
\keyword{may}
\keyword{model}
\keyword{network)}
\keyword{passed}
\keyword{restricted}
\keyword{returned}
\keyword{should}
\keyword{target.genes}
\keyword{the}
\keyword{those}
\keyword{to}
\keyword{used}
\keyword{whether}
