% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{plotDrugs}
\alias{plotDrugs}
\title{\code{plotDrugs} carries out the plotting mechanism for  a single network}
\usage{
plotDrugs(eset, drugList, genotype)
}
\arguments{
\item{eset}{expressionset}

\item{drugList}{list of drugs}
}
\description{
\code{plotDrugs} carries out the plotting mechanism for  a single network
}
