% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeResults.R
\name{plotDrugsAcrossData}
\alias{plotDrugsAcrossData}
\title{\code{plotDrugsAcrossData} Gets drugs from network and plots across input data}
\usage{
plotDrugsAcrossData(synId, tableId, genotype = "nf1 genotype")
}
\arguments{
\item{synId}{synapse id of stored network result}

\item{tableId}{of stored fendR results}
}
\value{
data frame of all go enrichment with adjusted p-value less than 0.1
}
\description{
\code{plotDrugsAcrossData} Gets drugs from network and plots across input data
}
