% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{predict.fendR}
\alias{predict.fendR}
\title{Allows for prediction from one fendR object to another}
\usage{
\method{predict}{fendR}(fObj, otherObj, modelCall = "lm", modelArgs = c(),
  predictArgs = c(), numDrugs = 10, extraLib = NA)
}
\arguments{
\item{modelCall}{A string representing the command to call to do the prediction}

\item{modelArgs}{A list of arguments to pass into the modelCall}

\item{numDrugs}{Number of drugs to limit to}

\item{extraLib}{Name of library that requires loading}

\item{object}{a fendrObject from which to build a model}

\item{otherObject}{a second fendRObject to predict to}
}
\value{
A data frame comparing the true values of the 'other' fendR object as well #' as the predicted models under a basic model and an engineered model
}
\description{
Evaluate how well one fendRized dataset predicts another
}
