% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{reduceGenes.fendR}
\alias{reduceGenes.fendR}
\title{Predict from one object to another
Reduce model to only consider a subset of genes}
\usage{
\method{reduceGenes}{fendR}(object, gene.list)
}
\description{
Predict from one object to another
Reduce model to only consider a subset of genes
}
