% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{reducePhenos.fendR}
\alias{reducePhenos.fendR}
\title{Reduce model to only consider a set of phenotypes}
\usage{
\method{reducePhenos}{fendR}(object, phenos)
}
\description{
Reduce model to only consider a set of phenotypes
}
