% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{reduceSamples.fendR}
\alias{reduceSamples.fendR}
\title{Reduce mode to only consider a set of samples}
\usage{
\method{reduceSamples}{fendR}(object, samp.list)
}
\description{
Reduce mode to only consider a set of samples
}
