% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{removeSampleFromObject}
\alias{removeSampleFromObject}
\title{Helper function to remove feature data}
\usage{
removeSampleFromObject(obj, sampleName)
}
\arguments{
\item{fendRObj}{The object class}
}
\value{
updated object
}
\description{
Un-exported for now
}
\keyword{leaveOneOut}
