% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{removeSampleOrPhenoFromObject}
\alias{removeSampleOrPhenoFromObject}
\title{Helper function to remove feature data or sample data from any fendR object}
\usage{
removeSampleOrPhenoFromObject(obj, sampleName = "", phenoName = "")
}
\arguments{
\item{fendRObj}{The object class}
}
\value{
updated object
}
\description{
Un-exported for now
}
\keyword{leaveOneOut}
