% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\name{runForestWithParams}
\alias{runForestWithParams}
\title{Run Forest using python code
\code{runForestWithParams} runs python code and returns graph objects}
\usage{
runForestWithParams(forestPath, mu, beta, w, prizeFileName, netFileName,
  depth = 10)
}
\arguments{
\item{object}{That contains a data frame and network}
}
\value{
list of graph objects
}
\description{
Run Forest using python code
\code{runForestWithParams} runs python code and returns graph objects
}
