% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionToViper.R
\name{runViperOnDrugs}
\alias{runViperOnDrugs}
\title{\code{runViperOnDrugs} takes an expression set and a list of drugs and identifies a viper list of proteins that explain differential expression}
\usage{
runViperOnDrugs(eset, drugs = c("BRD-K11533227", "dacarbazine"))
}
\arguments{
\item{eset}{is expression set}

\item{drugs}{is a list of drug names}
}
\value{
viper object
}
\description{
\code{runViperOnDrugs} takes an expression set and a list of drugs and identifies a viper list of proteins that explain differential expression
}
