% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionToViper.R
\name{runViperOnDset}
\alias{runViperOnDset}
\title{\code{runViperOnDset} takes an expression set and runs viper with all aracne
# networks}
\usage{
runViperOnDset(eset)
}
\arguments{
\item{eset}{is expression set}
}
\value{
viper object
}
\description{
\code{runViperOnDset} takes an expression set and runs viper with all aracne
# networks
}
