% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFendR.R
\name{scoreDataFromModel.basicFendR}
\alias{scoreDataFromModel.basicFendR}
\title{\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set}
\usage{
\method{scoreDataFromModel}{basicFendR}(object)
}
\value{
list of scores for each of the columns of the unseen feature data frame
}
\description{
\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set
}
