% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestFendR.R
\name{scoreDataFromModel.forestFendR}
\alias{scoreDataFromModel.forestFendR}
\title{\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set}
\usage{
\method{scoreDataFromModel}{forestFendR}(object)
}
\value{
list of scores for each of the columns of the unseen feature data frame
}
\description{
\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set
}
