% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{scoreDataFromModel}
\alias{scoreDataFromModel}
\title{\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set}
\usage{
scoreDataFromModel(model, unseenFeatures)
}
\value{
list of scores for each of the columns of the unseen feature data frame
}
\description{
\code{scoreDataFromModel} takes the new model and predicts a phenotype from an input set
}
