% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{translateMatrixIdentifiers}
\alias{translateMatrixIdentifiers}
\title{translateMatrixIdentifiers}
\usage{
translateMatrixIdentifiers(matrix, translationTable, agg.fun = "mean",
  debug = FALSE)
}
\arguments{
\item{matrix}{A symmetric matrix}

\item{translationTable}{A data.frame that maps from names in column 'from' to names in column 'to'}

\item{agg.fun}{A string name of a function that will be used to aggregrate multiple columns (from) into a single column (to).  Currently: mean, min, or max.}
}
\value{
A symmetric matrix with columns and rows translated (including dropped or duplicated) according to the translationTable.
}
\description{
\code{translateMatrixIdentifiers} translates the column and row identifers based on a translation table.  This may drop and/or duplicate columns and rows.
}
