% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varCompWrapperFunction.R
\name{varCompWrapperFunction}
\alias{varCompWrapperFunction}
\title{\code{varCompWrapperFunction} takes outcome and features}
\usage{
varCompWrapperFunction(outcome, features)
}
\description{
\code{varCompWrapperFunction} takes outcome and features
}
\keyword{phenotype}
