% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{createNewFeaturesFromNetwork}
\alias{createNewFeaturesFromNetwork}
\title{Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the gene-based measurements and alters their #' score using a network}
\usage{
createNewFeaturesFromNetwork(fendrObject)
}
\arguments{
\item{fendrObject}{That contains a data frame and network}
}
\value{
data.frame representing new gene by sample matrix that is augmented
}
\description{
Engineer Features from Network
\code{createNewFeaturesFromNetwork} takes the gene-based measurements and alters their #' score using a network
}
