% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{crossValidationCompare}
\alias{crossValidationCompare}
\title{Assess fendR and modeling using cross validation}
\usage{
crossValidationCompare(fendRObj, modelCall = "lm", modelArgs = list(),
  testPheno = c(), sampleIndependent = TRUE)
}
\arguments{
\item{fendRObj}{The object class}

\item{modelCall}{A String to call the model and predict}

\item{modelArgs}{a list of extra arguments to pass into do.call with model}

\item{testPheno}{a list of phenotypes to limit the test}

\item{sampleIndependent}{set to TRUE if your engineered features are independent of samples}
}
\value{
Not sure yet...
}
\description{
Use leave-one-out cross validation to assess the feature mapping and
modeling approach
}
\keyword{INCOMPLETE}
