% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{edgeList2matrix}
\alias{edgeList2matrix}
\title{edgeList2matrix convertion and write out}
\usage{
edgeList2matrix(elPath, outPath = NULL)
}
\arguments{
\item{elPath}{file path to edge list file with three colums, geneA, geneB and edge}

\item{outPath}{directory for output if NULL it will be same as input file}
}
\value{
NULL if successfull,error otherwise
}
\description{
\code{edgeList2matrix} loads in the target/phenotype data with the first column representing the drug of choice and the second represneting the gene
}
