% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{fendR}
\alias{fendR}
\title{An S3 class to represent a fendR predictive network algorithm}
\usage{
fendR(networkFile, featureData, phenoFeatureData, sampleOutcomeData,
  targetGenes = NULL, geneNorm = NA, responseNorm = NA)
}
\arguments{
\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{phenoFeatureData}{a data.frame where rows represent genes and columns represent a relationship between phenotype and gene}

\item{sampleOutcomeData}{a data.frame representing at least one column of phenotype and rows representing samples}

\item{network}{A network or path to network, depending on the underlying class}
}
\value{
a fendR object
}
\description{
An S3 class to represent a fendR predictive network algorithm
}
