% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{loadNetwork}
\alias{loadNetwork}
\title{Load network from file or path}
\usage{
loadNetwork(object)
}
\arguments{
\item{object}{A fendR object}

\item{testPheno}{A set of phenotypes to limit the analysis scope}
}
\value{
A fendR object with a data frame with three columns: Gene, Phenotype, NetworkScore
}
\description{
Takes the network handle and loads relevant features from network into iGraph object
}
\keyword{network}
