% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3FendR.R
\name{n3FendR}
\alias{n3FendR}
\title{An S3 class to represent the nearest-neighbor network implementation of the fendR predictive
network algorithm.}
\usage{
n3FendR(network, featureData, phenoFeatureData, sampleOutcomeData, target.genes,
  network.type)
}
\arguments{
\item{network}{the file name of a feather or big.matrix object representing a dense matrix}

\item{featureData}{a data.frame that contains rows representing genes and columns representing samples}

\item{phenoFeatureData}{a data.frame where rows represent genes and columns represent a relationship between phenotype and gene}

\item{sampleOutcomeData}{a data.frame representing at least one column of phenotype and rows representing samples}

\item{target.genes}{a vector of target gene names (a subset of those in featureData) that will be used to sparsify the network}

\item{network.type}{the on-disk representation of the network (either "feather" or "big.memory")}
}
\value{
n3FendR object
}
\description{
An S3 class to represent the nearest-neighbor network implementation of the fendR predictive
network algorithm.
}
