% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{plotModelResults}
\alias{plotModelResults}
\title{Visualirze fendR and modeling using cross validation}
\usage{
plotModelResults(modelingDataFrame)
}
\arguments{
\item{modelingDataFrame}{output of \code{crossValidationCompare}}
}
\value{
image
}
\description{
plots results of LOO CV code above
}
