% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fendR.R
\name{predict.fendR}
\alias{predict.fendR}
\title{Allows for prediction from one fendR object to another}
\usage{
\method{predict}{fendR}(fObj, otherObj, modelCall = "lm", modelArgs = c(),
  numDrugs = 10)
}
\arguments{
\item{object}{a fendrObject}
}
\value{
A response matrix to use for modeling with the formula 'Response~.'
}
\description{
Evaluate how well one fendRized dataset predicts another
}
